Rework emails test tools (#1583)
Wrap each Email case class in a EmailDefinitions object, to force us to always declare email examples for the admin tests tools
