+++ app/controllers/AdminController.scala
@@ -1,32 +1,19 @@
 package controllers
 
 import authentication.Authenticator
+import authentication.actions.UserAction.WithRole
 import cats.data.NonEmptyList
 import cats.implicits.toTraverseOps
 import config.EmailConfiguration
-import models.report.DetailInputValue.toDetailInputValue
 import models._
 import models.admin.ReportInputList
-import models.auth.AuthToken
-import models.company.Address
-import models.company.Company
 import models.event.Event
-import models.report.ReportFileOrigin.Consumer
-import models.report.reportfile.ReportFileId
-import models.report.Gender
-import models.report.Report
-import models.report.ReportFile
-import models.report.ReportFilter
-import models.report.ReportResponse
-import models.report.ReportResponseType
-import models.report.ReportStatus
-import models.report.ReportTag
-import models.report.ResponseDetails
-import models.report.WebsiteURL
+import models.report._
 import orchestrators.ReportFileOrchestrator
 import play.api.Logger
 import play.api.i18n.Lang
 import play.api.i18n.Messages
+import play.api.i18n.MessagesApi
 import play.api.i18n.MessagesImpl
 import play.api.libs.json.JsError
 import play.api.libs.json.Json
@@ -37,52 +24,27 @@ import repositories.companyaccess.CompanyAccessRepositoryInterface
 import repositories.event.EventRepositoryInterface
 import repositories.report.ReportRepositoryInterface
 import repositories.subscription.SubscriptionRepositoryInterface
-import services.Email.AdminAccessLink
-import services.Email.AdminProbeTriggered
-import services.Email.ConsumerProResponseNotification
-import services.Email.ConsumerProResponseNotificationOnAdminCompletion
-import services.Email.ConsumerReportAcknowledgment
-import services.Email.ConsumerReportClosedNoAction
-import services.Email.ConsumerReportClosedNoReading
-import services.Email.ConsumerReportReadByProNotification
-import services.Email.ConsumerValidateEmail
-import services.Email.DgccrfAgentAccessLink
-import services.Email.DgccrfDangerousProductReportNotification
-import services.Email.DgccrfInactiveAccount
-import services.Email.DgccrfReportNotification
-import services.Email.DgccrfValidateEmail
-import services.Email.ProCompaniesAccessesInvitations
-import services.Email.ProCompanyAccessInvitation
-import services.Email.ProNewCompaniesAccesses
-import services.Email.ProNewCompanyAccess
-import services.Email.ProNewReportNotification
-import services.Email.ProReportAssignedNotification
-import services.Email.ProReportReOpeningNotification
-import services.Email.ProReportsReadReminder
-import services.Email.ProReportsUnreadReminder
-import services.Email.ProResponseAcknowledgment
-import services.Email.ProResponseAcknowledgmentOnAdminCompletion
-import services.Email.ReportDeletionConfirmation
-import services.Email.ResetPassword
-import services.Email.UpdateEmailAddress
-import services.Email
-import services.MailService
 import services.PDFService
-import utils.Constants.ActionEvent.POST_ACCOUNT_ACTIVATION_DOC
+import services.emails.EmailDefinitionsConsumer._
+import services.emails.EmailDefinitionsDggcrf._
+import services.emails.EmailDefinitionsPro._
+import services.emails.EmailsExamplesUtils._
+import services.emails.BaseEmail
+import services.emails.EmailDefinition
+import services.emails.EmailDefinitions.allEmailDefinitions
+import services.emails.MailService
 import utils.Constants.ActionEvent.REPORT_PRO_RESPONSE
 import utils.Constants.ActionEvent
 import utils.Constants.EventType
 import utils._
-import authentication.actions.UserAction.WithRole
 
-import java.time.LocalDate
 import java.time.OffsetDateTime
-import java.time.Period
 import java.util.Locale
 import java.util.UUID
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 import scala.concurrent.duration._
+
 class AdminController(
     reportRepository: ReportRepositoryInterface,
     companyAccessRepository: CompanyAccessRepositoryInterface,
@@ -103,121 +65,8 @@ class AdminController(
   implicit val contactAddress: EmailAddress = emailConfiguration.contactAddress
   implicit val timeout: akka.util.Timeout   = 5.seconds
 
-  val dummyURL = java.net.URI.create("https://lien-test")
-
-  private def genReport = Report(
-    id = UUID.fromString("c1cbadb3-04d8-4765-9500-796e7c1f2a6c"),
-    gender = Some(Gender.Female),
-    category = "Test",
-    subcategories = List("test"),
-    details = List(toDetailInputValue("test")),
-    companyId = Some(UUID.randomUUID()),
-    companyName = Some("Dummy Inc."),
-    companyBrand = Some("Dummy Inc. Store"),
-    companyAddress = Address(Some("3 bis"), Some("Rue des exemples"), None, Some("13006"), Some("Douceville")),
-    companySiret = Some(SIRET("12345678912345")),
-    companyActivityCode = None,
-    websiteURL = WebsiteURL(None, None),
-    phone = None,
-    creationDate = OffsetDateTime.now(),
-    firstName = "John",
-    lastName = "Doe",
-    email = EmailAddress("john.doe@example.com"),
-    contactAgreement = true,
-    employeeConsumer = false,
-    status = ReportStatus.TraitementEnCours,
-    expirationDate = OffsetDateTime.now().plusDays(50),
-    influencer = None,
-    visibleToPro = true,
-    lang = Some(Locale.FRENCH),
-    barcodeProductId = None,
-    train = None,
-    station = None
-  )
-
-  private def genReportFile = ReportFile(
-    id = ReportFileId.generateId(),
-    reportId = Some(UUID.fromString("c1cbadb3-04d8-4765-9500-796e7c1f2a6c")),
-    creationDate = OffsetDateTime.now(),
-    filename = s"${UUID.randomUUID.toString}.png",
-    storageFilename = "String",
-    origin = Consumer,
-    avOutput = None
-  )
-
-  private def genReportResponse = ReportResponse(
-    responseType = ReportResponseType.ACCEPTED,
-    responseDetails = Some(ResponseDetails.REFUND),
-    otherResponseDetails = None,
-    consumerDetails = "",
-    dgccrfDetails = Some(""),
-    fileIds = Nil
-  )
-
-  private def genCompany = Company(
-    id = UUID.randomUUID,
-    siret = SIRET.fromUnsafe("12345678901234"),
-    creationDate = OffsetDateTime.now(),
-    name = "Test Entreprise",
-    address = Address(
-      number = Some("3"),
-      street = Some("rue des Champs"),
-      postalCode = Some("75015"),
-      city = Some("Paris"),
-      country = None
-    ),
-    activityCode = None,
-    isHeadOffice = true,
-    isOpen = true,
-    isPublic = true,
-    brand = Some("une super enseigne")
-  )
-
-  private def genCompanyList = List(genCompany, genCompany, genCompany)
-
-  private def genSiren = SIREN.fromSIRET(genCompany.siret)
-
-  private def genUser = User(
-    id = UUID.randomUUID,
-    password = "",
-    email = EmailAddress("text@example.com"),
-    firstName = "Jeanne",
-    lastName = "Dupont",
-    userRole = UserRole.Admin,
-    lastEmailValidation = None
-  )
-
-  private def genEvent =
-    Event(
-      UUID.randomUUID(),
-      None,
-      None,
-      None,
-      OffsetDateTime.now(),
-      EventType.PRO,
-      POST_ACCOUNT_ACTIVATION_DOC
-    )
-
-  private def genAuthToken =
-    AuthToken(UUID.randomUUID, UUID.randomUUID, OffsetDateTime.now().plusDays(10))
-
-  private def genSubscription = Subscription(
-    id = UUID.randomUUID,
-    userId = None,
-    email = None,
-    departments = List("75"),
-    countries = Nil,
-    withTags = Nil,
-    withoutTags = Nil,
-    categories = Nil,
-    sirets = Nil,
-    frequency = java.time.Period.ofDays(1)
-  )
-
-  case class EmailContent(subject: String, body: play.twirl.api.Html)
-
   implicit val messagesProvider: Messages = MessagesImpl(Lang(Locale.FRENCH), controllerComponents.messagesApi)
-  val availablePdfs = Seq[(String, Html)](
+  private val availablePdfs = Seq[(String, Html)](
     "accountActivation" -> views.html.pdfs.accountActivation(
       genCompany,
       Some(genReport.creationDate.toLocalDate),
@@ -262,331 +111,31 @@ class AdminController(
     )(messagesProvider)
   )
 
-  val availableEmails = Map[String, EmailAddress => Email](
-    // ======= Divers =======
-    "various.reset_password"       -> (recipient => ResetPassword(genUser.copy(email = recipient), genAuthToken)),
-    "various.update_email_address" -> (recipient => UpdateEmailAddress(recipient, dummyURL, daysBeforeExpiry = 2)),
-
-    // ======= Admin =======
-    "admin.access_link" -> (recipient => AdminAccessLink(recipient, dummyURL)),
-    "admin.probe_triggered" -> (recipient =>
-      AdminProbeTriggered(Seq(recipient), "Taux de schtroumpfs pas assez schtroumpfés", 0.2, "bas")
-    ),
-
-    // ======= DGCCRF =======
-    "dgccrf.access_link" ->
-      (DgccrfAgentAccessLink("DGCCRF")(_, frontRoute.dashboard.Agent.register(token = "abc"))),
-    "dgccrf.inactive_account_reminder" -> (recipient =>
-      DgccrfInactiveAccount(genUser.copy(email = recipient), Some(LocalDate.now().plusDays(90)))
-    ),
-    "dgccrf.report_dangerous_product_notification" -> (recipient =>
-      DgccrfDangerousProductReportNotification(Seq(recipient), genReport)
-    ),
-    "dgccrf.report_notif_dgccrf" -> (recipient =>
-      DgccrfReportNotification(
-        List(recipient),
-        genSubscription,
-        List(
-          (genReport, List(genReportFile)),
-          (genReport.copy(tags = List(ReportTag.ReponseConso)), List(genReportFile))
-        ),
-        LocalDate.now().minusDays(10)
-      )
-    ),
-    "dgccrf.validate_email" ->
-      (DgccrfValidateEmail(_, 7, frontRoute.dashboard.validateEmail(""))),
-
-    // ======= PRO =======
-    "pro.access_invitation" -> (recipient => ProCompanyAccessInvitation(recipient, genCompany, dummyURL, None)),
-    "pro.access_invitation_multiple_companies" -> (recipient =>
-      ProCompaniesAccessesInvitations(recipient, genCompanyList, genSiren, dummyURL)
-    ),
-    "pro.new_company_access"     -> (recipient => ProNewCompanyAccess(recipient, genCompany, None)),
-    "pro.new_companies_accesses" -> (recipient => ProNewCompaniesAccesses(recipient, genCompanyList, genSiren)),
-    "pro.report_ack_pro" -> (recipient =>
-      ProResponseAcknowledgment(genReport, genReportResponse, genUser.copy(email = recipient))
-    ),
-    "pro.report_ack_pro_on_admin_completion" -> (recipient =>
-      ProResponseAcknowledgmentOnAdminCompletion(genReport, List(genUser.copy(email = recipient), genUser, genUser))
-    ),
-    "pro.report_notification" -> (recipient => ProNewReportNotification(NonEmptyList.of(recipient), genReport)),
-    "pro.report_reopening_notification" -> (recipient => ProReportReOpeningNotification(List(recipient), genReport)),
-    "pro.reports_transmitted_reminder" -> (recipient => {
-      val report1 = genReport
-      val report2 = genReport.copy(companyId = report1.companyId)
-      val report3 = genReport.copy(companyId = report1.companyId, expirationDate = OffsetDateTime.now().plusDays(5))
-      ProReportsReadReminder(
-        List(recipient),
-        List(report1, report2, report3),
-        Period.ofDays(7)
-      )
-    }),
-    "pro.reports_unread_reminder" -> (recipient => {
-      val report1 = genReport
-      val report2 = genReport.copy(companyId = report1.companyId)
-      val report3 = genReport.copy(companyId = report1.companyId, expirationDate = OffsetDateTime.now().plusDays(5))
-
-      ProReportsUnreadReminder(
-        List(recipient),
-        List(report1, report2, report3),
-        Period.ofDays(7)
-      )
-    }),
-    "pro.report_assignement_to_other" -> (recipient =>
-      ProReportAssignedNotification(
-        report = genReport,
-        assigningUser = genUser,
-        assignedUser = genUser.copy(email = recipient)
-      )
-    ),
-    "pro.report_deletion_confirmation" -> (_ =>
-      ReportDeletionConfirmation(
-        genReport,
-        Some(genCompany),
-        controllerComponents.messagesApi
-      )
-    ),
-
-    // ======= CONSO =======
-    "consumer.report_ack" -> (recipient =>
-      ConsumerReportAcknowledgment(
-        genReport.copy(email = recipient),
-        Some(genCompany),
-        genEvent,
-        Nil,
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_ack_case_reponseconso" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(status = ReportStatus.NA, tags = List(ReportTag.ReponseConso), email = recipient),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_dispute" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(tags = List(ReportTag.LitigeContractuel), email = recipient),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_dangerous_product" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.TraitementEnCours,
-            tags = List(ReportTag.ProduitDangereux),
-            email = recipient
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_euro" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.NA,
-            companyAddress = Address(country = Some(Country.Italie)),
-            email = recipient
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_euro_and_dispute" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.NA,
-            tags = List(ReportTag.LitigeContractuel),
-            companyAddress = Address(country = Some(Country.Islande)),
-            email = recipient
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_andorre" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.NA,
-            companyAddress = Address(country = Some(Country.Andorre)),
-            email = recipient
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_andorre_and_dispute" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.NA,
-            tags = List(ReportTag.LitigeContractuel),
-            companyAddress = Address(country = Some(Country.Andorre)),
-            email = recipient
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_suisse" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.NA,
-            companyAddress = Address(country = Some(Country.Suisse)),
-            email = recipient
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_suisse_and_dispute" -> (recipient =>
-      ConsumerReportAcknowledgment(
-        genReport.copy(
-          status = ReportStatus.NA,
-          tags = List(ReportTag.LitigeContractuel),
-          companyAddress = Address(country = Some(Country.Suisse)),
-          email = recipient
-        ),
-        Some(genCompany),
-        genEvent,
-        Nil,
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_ack_case_compagnie_aerienne" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.NA,
-            email = recipient,
-            tags = List(ReportTag.CompagnieAerienne)
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_abroad_default" ->
-      (recipient =>
-        ConsumerReportAcknowledgment(
-          genReport.copy(
-            status = ReportStatus.NA,
-            companyAddress = Address(country = Some(Country.Bahamas)),
-            email = recipient
-          ),
-          Some(genCompany),
-          genEvent,
-          Nil,
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_ack_case_abroad_default_and_dispute" -> (recipient =>
-      ConsumerReportAcknowledgment(
-        genReport.copy(
-          status = ReportStatus.NA,
-          tags = List(ReportTag.LitigeContractuel),
-          companyAddress = Address(country = Some(Country.Bahamas)),
-          email = recipient
-        ),
-        Some(genCompany),
-        genEvent,
-        Nil,
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_transmitted" -> (recipient =>
-      ConsumerReportReadByProNotification(
-        genReport.copy(email = recipient),
-        Some(genCompany),
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_ack_pro_consumer" -> (recipient =>
-      ConsumerProResponseNotification(
-        genReport.copy(email = recipient),
-        genReportResponse,
-        Some(genCompany),
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_ack_pro_consumer_on_admin_completion" -> (recipient =>
-      ConsumerProResponseNotificationOnAdminCompletion(
-        genReport.copy(email = recipient),
-        Some(genCompany),
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_closed_no_reading" -> (recipient =>
-      ConsumerReportClosedNoReading(
-        genReport.copy(email = recipient),
-        Some(genCompany),
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_closed_no_reading_case_dispute" ->
-      (recipient =>
-        ConsumerReportClosedNoReading(
-          genReport.copy(email = recipient, tags = List(ReportTag.LitigeContractuel)),
-          Some(genCompany),
-          controllerComponents.messagesApi
-        )
-      ),
-    "consumer.report_closed_no_action" -> (recipient =>
-      ConsumerReportClosedNoAction(
-        genReport.copy(email = recipient),
-        Some(genCompany),
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.report_closed_no_action_case_dispute" -> (recipient =>
-      ConsumerReportClosedNoAction(
-        genReport.copy(email = recipient, tags = List(ReportTag.LitigeContractuel)),
-        Some(genCompany),
-        controllerComponents.messagesApi
-      )
-    ),
-    "consumer.validate_email" -> (recipient =>
-      ConsumerValidateEmail(EmailValidation(email = recipient), None, controllerComponents.messagesApi)
-    )
-  )
+  private val allEmailExamples: Seq[(String, (EmailAddress, MessagesApi) => BaseEmail)] =
+    allEmailDefinitions.flatMap(readExamplesWithFullKey)
 
   def getEmailCodes = SecuredAction.andThen(WithRole(UserRole.Admin)).async { _ =>
-    Future(Ok(Json.toJson(availableEmails.keys)))
+    val keys = allEmailExamples.map(_._1)
+    Future(Ok(Json.toJson(keys)))
   }
   def sendTestEmail(templateRef: String, to: String) = SecuredAction.andThen(WithRole(UserRole.Admin)).async { _ =>
-    Future(
-      availableEmails
-        .get(templateRef)
-        .map(e => mailService.send(e(EmailAddress(to))))
-        .map(_ => Ok)
-        .getOrElse(NotFound)
-    )
+    val maybeEmail = allEmailExamples
+      .find(_._1 == templateRef)
+      .map(_._2(EmailAddress(to), controllerComponents.messagesApi))
+    maybeEmail match {
+      case Some(email) =>
+        mailService.send(email).map(_ => Ok)
+      case None =>
+        Future.successful(NotFound)
+    }
   }
+
+  private def readExamplesWithFullKey(
+      emailDefinition: EmailDefinition
+  ): Seq[(String, (EmailAddress, MessagesApi) => BaseEmail)] =
+    emailDefinition.examples.map { case (key, fn) =>
+      s"${emailDefinition.category.toString.toLowerCase}.$key" -> fn
+    }
 
   def getPdfCodes = SecuredAction.andThen(WithRole(UserRole.Admin)).async { _ =>
     Future(Ok(Json.toJson(availablePdfs.map(_._1))))
@@ -619,7 +168,7 @@ class AdminController(
       _ <- filteredEvents.map { case (report, responseEvent) =>
         val maybeCompany = report.companyId.flatMap(companyId => companies.find(_.id == companyId))
         mailService.send(
-          ConsumerProResponseNotification(
+          ConsumerProResponseNotification.Email(
             report,
             responseEvent.details.as[ReportResponse],
             maybeCompany,
@@ -653,7 +202,7 @@ class AdminController(
       }.sequence
       _ <- reportAndEmailList.map {
         case (report, Some(adminsEmails)) =>
-          mailService.send(ProNewReportNotification(adminsEmails, report))
+          mailService.send(ProNewReportNotification.Email(adminsEmails, report))
         case (report, None) =>
           logger.debug(s"Not sending email for report ${report.id}, no admin found")
           Future.unit
@@ -683,7 +232,7 @@ class AdminController(
         event match {
           case Some(evt) =>
             Some(
-              ConsumerReportAcknowledgment(
+              ConsumerReportAcknowledgment.Email(
                 report,
                 maybeCompany,
                 evt,
@@ -720,7 +269,7 @@ class AdminController(
       } else Future(Seq())
     _ <-
       if (ddEmails.nonEmpty) {
-        mailService.send(DgccrfDangerousProductReportNotification(ddEmails, report))
+        mailService.send(DgccrfDangerousProductReportNotification.Email(ddEmails, report))
       } else {
         Future.unit
       }
