+++ app/orchestrators/ReportAssignmentOrchestrator.scala
@@ -11,8 +11,8 @@ import play.api.Logger
 import repositories.event.EventRepositoryInterface
 import repositories.reportmetadata.ReportMetadataRepositoryInterface
 import repositories.user.UserRepositoryInterface
-import services.Email.ProReportAssignedNotification
-import services.MailService
+import services.emails.EmailDefinitionsPro.ProReportAssignedNotification
+import services.emails.MailService
 import utils.Constants
 
 import java.time.OffsetDateTime
@@ -42,7 +42,10 @@ class ReportAssignmentOrchestrator(
       _                       <- createAssignmentEvent(reportWithMetadata.report, assigningUser, newAssignedUser)
       _ <-
         if (assigningToSelf) Future.unit
-        else mailService.send(ProReportAssignedNotification(reportWithMetadata.report, assigningUser, newAssignedUser))
+        else
+          mailService.send(
+            ProReportAssignedNotification.Email(reportWithMetadata.report, assigningUser, newAssignedUser)
+          )
     } yield newAssignedUser
   }
