--- app/services/MailServiceInterface.scala
@@ -1,6 +1,0 @@
-package services
-import scala.concurrent.Future
-
-trait MailServiceInterface {
-  def send(email: Email): Future[Unit]
-}
