+++ app/services/emails/MailServiceInterface.scala
@@ -1,0 +1,7 @@
+package services.emails
+
+import scala.concurrent.Future
+
+trait MailServiceInterface {
+  def send(email: BaseEmail): Future[Unit]
+}
