+++ app/tasks/account/InactiveDgccrfAccountReminderTask.scala
@@ -7,8 +7,8 @@ import play.api.Logger
 import play.api.libs.json.Json
 import repositories.event.EventRepositoryInterface
 import repositories.user.UserRepositoryInterface
-import services.Email.DgccrfInactiveAccount
-import services.MailService
+import services.emails.EmailDefinitionsDggcrf.DgccrfInactiveAccount
+import services.emails.MailService
 import utils.Constants.ActionEvent
 import utils.Constants.EventType
 import utils.Logs.RichLogger
@@ -57,7 +57,7 @@ class InactiveDgccrfAccountReminderTask(
     val now            = OffsetDateTime.now()
     val expirationDate = user.lastEmailValidation.map(_.plus(inactivePeriod))
     for {
-      _ <- mailService.send(DgccrfInactiveAccount(user, expirationDate.map(_.toLocalDate)))
+      _ <- mailService.send(DgccrfInactiveAccount.Email(user, expirationDate.map(_.toLocalDate)))
       _ <- eventRepository.create(
         Event(
           UUID.randomUUID(),
