+++ app/views/mails/dgccrf/reportDangerousProductNotification.scala.html
@@ -77,7 +77,7 @@
         Pour consulter l'intégralité du signalement de produit dangereux, notamment les pièces jointes :
     </p>
     <ul>
-        <li>Vous devez suivre la procédure mentionner à l'instruction IN-2D-PIL-004</li>
+        <li>Vous devez suivre la procédure mentionnée à l'instruction IN-2D-PIL-004</li>
         <li>si vous possédez un compte sur SignalConso, connectez-vous sur <a href="@frontRoute.dashboard.login">@frontRoute.dashboard.login</a></li>
         <li>sinon, vous pouvez demander la création d'un compte en envoyant un mail avec vos nom, prénom et service à l'adresse <a href="mailto:@contactAddress">@contactAddress</a></li>
     </ul>
